@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <AssemblyName>LinqIntegrated</AssemblyName>
     <OutputType>Exe</OutputType>
     <PackageId>LinqIntegrated</PackageId>
