@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <AssemblyName>LinqIntro</AssemblyName>
     <OutputType>Exe</OutputType>
     <PackageId>LinqIntro</PackageId>    
